requirements.txt
click==8.0.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
mysql-connector-python
Flask-mysql==1.5.2